LINECHATBOT_RPLSMS/requirements.txt
line-bot-sdk
Flask
SQLAlchemy
PyMySQL
flask-sqlalchemy
#flask-mysql
psycopg2
Flask-Migrate